refactor(harness): move cli method listing
